Added init function and fixed file paths
