+++ utils/generateMarkdown.js
@@ -48,7 +48,7 @@ function renderLicenseBadge(license) {
       break;
 
   }
-}
+};
 
 // TODO: Create a function that returns the license link
 // If there is no license, return an empty string
@@ -101,7 +101,7 @@ function renderLicenseLink(license) {
 
   }
 
-}
+};
 
 // TODO: Create a function that returns the license section of README
 // If there is no license, return an empty string
@@ -172,7 +172,7 @@ function renderLicenseSection(license) {
 
   }
 
-}
+};
 
 // TODO: Create a function to generate markdown for README
 function generateMarkdown(data) {
@@ -195,6 +195,6 @@ function generateMarkdown(data) {
   renderLicenseBadge();
 `;
 return output;
-}
+};
 
 module.exports = generateMarkdown;
